kodi: don't enable webserver by default
Kodi defaults to require authentication when the webserver is
enabled and as no password is set this results in an annoying
error box on fresh installations.

As neither disabling authentication or pre-setting a well known
password are great options just stop enabling the webserver via
appliance.xml. Users can enable it and choose authentication options
if they like via settings.

Signed-off-by: Matthias Reichl <hias@horus.com>