linux (RPi4): disable legacy framebuffer and dmaengine drivers
The legacy drivers are not used and this frees up one of the
scarce DMA channels.

Keep the legacy drivers enabled on RPi and RPi2 for now, they
don't suffer from the DMA channel shortage and it may help downstream
projects which still do RPi builds using the legacy graphics stack.

Signed-off-by: Matthias Reichl <hias@horus.com>