WIP: gcc: libgcc/libgcov.h: bootstrap failure due to missing #include <sys/mman.h>
Workaround with bug:
- https://gcc.gnu.org/bugzilla/show_bug.cgi?id=100289

Issue is also present with cross-compiler (11.2) on FreeBSD:

Can confirm this bug when building an AVR cross-compiler (11.2) on
FreeBSD. To get it working, I'm now patching it to #undef
HAVE_SYS_MMAN_H in libgcov.h before starting.

This is a WIP - if this patch / workaround is to be used - it should
only be used in the bootstrap stage of the compiler build.